separate add block function into a hook
